+++ Expo/AuthProvider.tsx
@@ -6,7 +6,7 @@ import React, {
   ReactNode,
 } from "react";
 import { onAuthStateChanged, User } from "firebase/auth";
-import { auth } from "../../firebaseModel";
+import { auth } from "./firebaseModel";
 
 type AuthContextType = {
   user: User | null;
